backend/requirements.txt
starlette==0.22.0
uvicorn==0.20.0
python-multipart==0.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

